fix: resolve DEX class collisions between app/ and shared-ui/
When app/ and shared-ui/ have top-level Kotlin files with the same FQN,
D8 silently deduplicates them at build time, keeping the library (shared-ui)
version. This caused NoSuchMethodError crashes at runtime because the
shared-ui versions have different signatures (no Android-specific params).

Fixes applied:
- MainScreen.kt → renamed to AndroidAppShell.kt + function renamed to
  AndroidAppShell (produces AndroidAppShellKt vs shared-ui's MainScreenKt)
- AppTopBar.kt → @file:JvmName("AndroidAppTopBar")
- BottomNavItem.kt → @file:JvmName("AndroidBottomNavItem")
- BottomNavigationBar.kt → @file:JvmName("AndroidBottomNavigationBar")
- app/Route.kt (sealed class, can't use JvmName) → deleted; merged
  PresidentTeamDetail and PendingTeamAssignment into shared-ui's Route.kt
