+++ app/src/main/java/com/jesuslcorominas/teamflowmanager/ui/components/bottomnav/BottomNavItem.kt
@@ -1,3 +1,5 @@
+@file:JvmName("AndroidBottomNavItem")
+
 package com.jesuslcorominas.teamflowmanager.ui.components.bottomnav
 
 import androidx.compose.animation.core.Spring
