+++ app/src/main/java/com/jesuslcorominas/teamflowmanager/ui/components/topbar/AppTopBar.kt
@@ -1,3 +1,5 @@
+@file:JvmName("AndroidAppTopBar")
+
 package com.jesuslcorominas.teamflowmanager.ui.components.topbar
 
 import androidx.activity.compose.BackHandler
