+++ app/src/main/java/com/jesuslcorominas/teamflowmanager/ui/main/MainActivity.kt
@@ -61,7 +61,7 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             TFMAppTheme {
-                MainScreen(
+                AndroidAppShell(
                     pendingNavigation = pendingNavigation,
                 )
             }
