+++ app/src/main/java/com/jesuslcorominas/teamflowmanager/ui/navigation/BottomNavigationBar.kt
@@ -1,3 +1,5 @@
+@file:JvmName("AndroidBottomNavigationBar")
+
 package com.jesuslcorominas.teamflowmanager.ui.navigation
 
 import androidx.compose.foundation.layout.WindowInsets
